% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDHtoLatLong.R
\name{RDHtoLatLong}
\alias{RDHtoLatLong}
\title{Converte 'Rijksdriehoek' (RDH) Coordinates to Lat and Long (WGS84)}
\usage{
RDHtoLatLong(x, y)
}
\arguments{
\item{x}{x vector of x-coordinates in rijksdriehoek system.}

\item{y}{y vector of y-coordinates in rijksdriehoek system.}
}
\value{
A vector of length 2 with latitude and longitude.
}
\description{
Converte 'Rijksdriehoek' (RDH) Coordinates to Lat and Long (WGS84)
}
\examples{
# Converting the coordinates of the city of Amsterdam in RDH to Lat Long
x <- 121687
y <- 487484

RDHtoLatLong(x, y)

}
